Load scene data from FBX tree
Read the scene tree and assign entities for each node in the hierarchy,
assigning FbxMesh to their proper place in it.

The current implementation still has major issues,
visible in the Paris Bistro scene from Amazon research.

For example the sewer covers, one of the plant pots, the bistro door
and the light garlands are not moved properly to their right place.

However, now the scene is the right side up, which makes it much easier
to deal with.